Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/windowclient-navigate.https-expected.txt


FAIL normal assert_equals: expected "https://localhost:9443/service-workers/service-worker/resources/blank.html?navigate" but got "NotSupportedError"
FAIL blank url assert_equals: expected "https://localhost:9443/service-workers/service-worker/resources/windowclient-navigate-worker.js" but got "NotSupportedError"
FAIL in scope but not controlled test on installing worker assert_equals: expected "TypeError" but got "NotSupportedError"
FAIL in scope but not controlled test on active worker assert_equals: expected "TypeError" but got "NotSupportedError"
FAIL out of scope assert_equals: expected "TypeError" but got "NotSupportedError"
FAIL cross orgin url assert_equals: expected (object) null but got (string) "NotSupportedError"
FAIL invalid url (http://[example.com]) assert_equals: expected "TypeError" but got "NotSupportedError"
FAIL invalid url (view-source://example.com) assert_equals: expected "TypeError" but got "NotSupportedError"
FAIL invalid url (file:///) assert_equals: expected "TypeError" but got "NotSupportedError"
FAIL invalid url (about:blank) assert_equals: expected "TypeError" but got "NotSupportedError"
